+++ jira/tests/__init__.py
@@ -1,6 +1,0 @@
-"""
-Copyright start
-MIT License
-Copyright (c) 2025 Fortinet Inc
-Copyright end
-"""